# frontend/swagger-file.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Documentation",
    "description": "Description of your API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "fetchAllTransactions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDTOTransactionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "updateTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "deleteTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exchange-rates": {
      "get": {
        "tags": [
          "exchange-rate-controller"
        ],
        "operationId": "fetchAllExchangeRates",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDTOExchangeRateDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "exchange-rate-controller"
        ],
        "operationId": "updateExchangeRate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeRateDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exchange-rate-controller"
        ],
        "operationId": "createExchangeRate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeRateDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "exchange-rate-controller"
        ],
        "operationId": "deleteExchangeRate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeRateDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/test/transactions": {
      "post": {
        "tags": [
          "test-controller"
        ],
        "operationId": "createTransaction_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transactions/{id}": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "fetchTransactionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exchange-rates/{id}": {
      "get": {
        "tags": [
          "exchange-rate-controller"
        ],
        "operationId": "fetchExchangeRateById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeRateDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/customers/transactions/{customerId}": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "fetchAllTransactionsByCustomerId",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "woreda": {
            "type": "string"
          },
          "kebele": {
            "type": "string"
          },
          "subCity": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "BankAccount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountNumber": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "transactions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "branchCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "users": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "Currency": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "currencyCode": {
            "maxLength": 3,
            "minLength": 0,
            "type": "string"
          },
          "country": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "middleName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "lastName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "identificationType": {
            "type": "string",
            "enum": [
              "PASSPORT",
              "NATIONAL_IDENTIFICATION",
              "DRIVER_LICENSE"
            ]
          },
          "identificationNumber": {
            "type": "string"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "bankAccounts": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BankAccount"
            }
          },
          "transactions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "ExchangeRate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "baseCurrency": {
            "$ref": "#/components/schemas/Currency"
          },
          "targetCurrency": {
            "$ref": "#/components/schemas/Currency"
          },
          "rate": {
            "type": "number",
            "format": "double"
          },
          "transactionType": {
            "type": "string"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "module": {
            "type": "string",
            "enum": [
              "USER_MANAGEMENT",
              "TRANSACTION_MANAGEMENT",
              "REPORT_MANAGEMENT",
              "EXCHANGE_RATES_MANAGEMENT"
            ]
          },
          "access": {
            "type": "string",
            "enum": [
              "READ",
              "WRITE"
            ]
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "enum": [
              "ROLE_ADMIN",
              "ROLE_USER",
              "ROLE_MODERATOR"
            ]
          },
          "permissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "ticketNumber": {
            "type": "integer",
            "format": "int64"
          },
          "permitNumber": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "PURCHASE",
              "SELL"
            ]
          },
          "transactionStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "CANCELLED",
              "COMPLETED"
            ]
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "paymentMethodType": {
            "type": "string",
            "enum": [
              "CASH",
              "BANK_TRANSFER",
              "CARD"
            ]
          },
          "bankAccount": {
            "$ref": "#/components/schemas/BankAccount"
          },
          "transactionItems": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionItem"
            }
          },
          "travelDocument": {
            "$ref": "#/components/schemas/TravelDocument"
          }
        }
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "ticketNumber": {
            "type": "integer",
            "format": "int64"
          },
          "permitNumber": {
            "type": "string"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "PURCHASE",
              "SELL"
            ]
          },
          "transactionStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "CANCELLED",
              "COMPLETED"
            ]
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "transactionItems": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionItem"
            }
          },
          "travelDocument": {
            "$ref": "#/components/schemas/TravelDocument"
          },
          "paymentMethodType": {
            "type": "string",
            "enum": [
              "CASH",
              "BANK_TRANSFER",
              "CARD"
            ]
          },
          "bankAccount": {
            "$ref": "#/components/schemas/BankAccount"
          }
        }
      },
      "TransactionItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "exchangeRate": {
            "$ref": "#/components/schemas/ExchangeRate"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          },
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          }
        }
      },
      "TravelDocument": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "passportNumber": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "nationality": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "airTicket": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "residenceId": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "countryOfResidence": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "residenceExpires": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "visaNumber": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "visaType": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "visaIssuedBy": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "visaIssuedOn": {
            "type": "string",
            "format": "date-time"
          },
          "visaValidity": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "purposeOfTravel": {
            "type": "string"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          },
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          }
        }
      },
      "User": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "DELETED"
            ]
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "branch": {
            "$ref": "#/components/schemas/Branch"
          },
          "customers": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            }
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ExchangeRateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "baseCurrency": {
            "$ref": "#/components/schemas/Currency"
          },
          "targetCurrency": {
            "$ref": "#/components/schemas/Currency"
          },
          "rate": {
            "type": "number",
            "format": "double"
          },
          "transactionType": {
            "type": "string"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginatedResponseDTOTransactionDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDto"
            }
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PaginatedResponseDTOExchangeRateDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExchangeRateDto"
            }
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}